backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
alembic==1.13.1